On branch main Your branch is ahead of 'origin/main' by 1 commit.   (use "git push" to publish your local commits)
Changes to be committed:
	modified:   app/Http/Controllers/CourseController.php
	modified:   app/Http/Controllers/StudentController.php

+++ app/Http/Controllers/CourseController.php
@@ -195,11 +195,17 @@ class CourseController extends Controller
      */
     public function destroy(Course $course)
     {
-        // eliminamos tambien el archivo
+        $course->load('inscriptions');
+        if ($course->inscriptions->count() > 0) {
+            return response()->json([
+                "success" => false,
+                "message" => "No se puede eliminar el curso porque tiene inscripciones asociadas",
+                "data" => $course
+            ]);
+        }
+
         Storage::delete($this->IMAGE_PATH . "/" . $course->image);
-
         $course->delete();
-
         return response()->json([
             "success" => true,
             "message" => "Recurso eliminado",
